// src/types/booking.ts
export interface BookingData {
    organizerId: string;
    userId: string;
    eventName: string;
    userName: string;
    contactNumber: string;
    email: string;
    bookingTime: string;
    eventType: string;
    bookedAt: string;
    bookingDate: string;
    bookingId: string;
    status: string
}